+++ modules/hosts/desktop/default.nix
@@ -88,9 +88,8 @@
     ];
   };
 
-  # Install Chromium.
-  programs.chromium.enable = true;
   programs.zsh.enable = true;
+  features.browser.enable = true;
   features.python.enable = true;
   features.nodejs.enable = true;
   # >>> gpu-doctor graphics